SNTAB-86 change games logic

+++ handlers/gamesHandlers/gameHandlers_test.go
@@ -2,6 +2,7 @@ package gameshandlers_test
 
 import (
 	"encoding/json"
+	"fmt"
 	"net/http"
 	"strconv"
 	"testing"
@@ -9,7 +10,6 @@ import (
 	gameshandlers "github.com/CardozoCasariegoLuciano/StudyNotes-backend/handlers/gamesHandlers"
 	"github.com/CardozoCasariegoLuciano/StudyNotes-backend/handlers/responses"
 	"github.com/CardozoCasariegoLuciano/StudyNotes-backend/helpers/environment"
-	"github.com/CardozoCasariegoLuciano/StudyNotes-backend/helpers/roles"
 	testtools "github.com/CardozoCasariegoLuciano/StudyNotes-backend/helpers/testTools"
 	mock_models "github.com/CardozoCasariegoLuciano/StudyNotes-backend/helpers/testTools/mocks"
 	dbmodels "github.com/CardozoCasariegoLuciano/StudyNotes-backend/models/dbModels"
@@ -27,21 +27,25 @@ func Test_AllGames_HaveGames(t *testing.T) {
 	mockUserRepo := mock_models.NewMockIstorage(ctrl)
 	game := gameshandlers.NewGame(mockUserRepo)
 
+	userOwner := 100
+
 	expectedGames := &dbmodels.Games{
 		dbmodels.Game{
 			Title:       "test",
 			Description: "test",
+			UserID:      userOwner,
 		},
 		dbmodels.Game{
 			Title:       "test2",
 			Description: "test2",
+			UserID:      userOwner,
 		},
 	}
 
 	mockUserRepo.
 		EXPECT().
-		GetAll(gomock.AssignableToTypeOf(&dbmodels.Games{})).
-		Do(func(games *dbmodels.Games) {
+		GetAllGames(gomock.Eq(userOwner), gomock.AssignableToTypeOf(&dbmodels.Games{})).
+		Do(func(userOwner int, games *dbmodels.Games) {
 			*games = *expectedGames
 		}).
 		Times(1)
@@ -50,11 +54,13 @@ func Test_AllGames_HaveGames(t *testing.T) {
 	testCases := []struct {
 		name            string
 		expectedCode    int
+		userID          int
 		expectedResonse responses.Response
 	}{
 		{
-			name:         "When have elements",
+			name:         "when user is owner",
 			expectedCode: http.StatusOK,
+			userID:       userOwner,
 			expectedResonse: responses.Response{
 				MessageType: "OK",
 				Message:     "All games",
@@ -65,9 +71,10 @@ func Test_AllGames_HaveGames(t *testing.T) {
 
 	for _, tc := range testCases {
 		testConfig := testtools.InitTestConfig{
-			Path:       "/api/v1" + "/games",
-			Method:     http.MethodGet,
-			ApplyToken: false,
+			Path:        "/api/v1" + "/games",
+			Method:      http.MethodGet,
+			ApplyToken:  true,
+			TokenUserID: tc.userID,
 		}
 		testData := testtools.SetGenericTestData(&testConfig)
 		context := *testData.Context
@@ -79,6 +86,7 @@ func Test_AllGames_HaveGames(t *testing.T) {
 
 			resp := responses.Response{}
 			err = json.Unmarshal(writer.Body.Bytes(), &resp)
+
 			assert.NoError(t, err)
 
 			//Test Cases
@@ -97,21 +105,25 @@ func Test_AllGames_EmptyGames(t *testing.T) {
 	mockUserRepo := mock_models.NewMockIstorage(ctrl)
 	game := gameshandlers.NewGame(mockUserRepo)
 
+	userID := 100
+
 	mockUserRepo.
 		EXPECT().
-		GetAll(gomock.AssignableToTypeOf(&dbmodels.Games{})).
+		GetAllGames(gomock.Eq(userID), gomock.AssignableToTypeOf(&dbmodels.Games{})).
 		Return(nil).
 		Times(1)
 
 	environment.SetTestEnvirontment()
 	testCases := []struct {
 		name            string
+		userID          int
 		expectedCode    int
 		expectedResonse responses.Response
 	}{
 		{
 			name:         "When no have elements",
 			expectedCode: http.StatusOK,
+			userID:       100,
 			expectedResonse: responses.Response{
 				MessageType: "OK",
 				Message:     "All games",
@@ -122,9 +134,10 @@ func Test_AllGames_EmptyGames(t *testing.T) {
 
 	for _, tc := range testCases {
 		testConfig := testtools.InitTestConfig{
-			Path:       "/api/v1" + "/games",
-			Method:     http.MethodGet,
-			ApplyToken: false,
+			Path:        "/api/v1" + "/games",
+			Method:      http.MethodGet,
+			ApplyToken:  true,
+			TokenUserID: tc.userID,
 		}
 		testData := testtools.SetGenericTestData(&testConfig)
 		context := *testData.Context
@@ -138,7 +151,7 @@ func Test_AllGames_EmptyGames(t *testing.T) {
 			err = json.Unmarshal(writer.Body.Bytes(), &resp)
 			assert.NoError(t, err)
 
-			//Test Cases
+			// Test Cases
 			assert.Equal(t, tc.expectedCode, writer.Code)
 			assert.Equal(t, tc.expectedResonse.Data, resp.Data)
 			assert.Equal(t, tc.expectedResonse.MessageType, resp.MessageType)
@@ -155,10 +168,12 @@ func Test_GameByID(t *testing.T) {
 	mockUserRepo := mock_models.NewMockIstorage(ctrl)
 	game := gameshandlers.NewGame(mockUserRepo)
 
+	userID := 100
+
 	foundID := 1
 	mockUserRepo.
 		EXPECT().
-		GetById(gomock.Eq(foundID), gomock.AssignableToTypeOf(&dbmodels.Game{})).
+		GetGameById(gomock.Eq(userID), gomock.Eq(foundID), gomock.AssignableToTypeOf(&dbmodels.Game{})).
 		Return(&gorm.DB{RowsAffected: 1}).
 		Times(1)
 
@@ -166,30 +181,29 @@ func Test_GameByID(t *testing.T) {
 	testCases := []struct {
 		name            string
 		gameID          string
-		userRole        string
+		userID          int
 		expectedCode    int
 		expectedResonse responses.Response
 	}{
 		{
 			name:         "Must show the game",
 			gameID:       strconv.Itoa(foundID),
-			userRole:     roles.ADMIN,
+			userID:       userID,
 			expectedCode: http.StatusOK,
 			expectedResonse: responses.Response{
 				MessageType: "OK",
 				Message:     "Game Selected",
-				Data:        nil,
 			},
 		},
 	}
 
 	for _, tc := range testCases {
 		testConfig := testtools.InitTestConfig{
-			Path:          "/api/v1" + "/games",
-			Method:        http.MethodGet,
-			ReqBody:       nil,
-			ApplyToken:    true,
-			TokenUserRole: tc.userRole,
+			Path:        "/api/v1" + "/games",
+			Method:      http.MethodGet,
+			ReqBody:     nil,
+			ApplyToken:  true,
+			TokenUserID: userID,
 		}
 		testData := testtools.SetGenericTestData(&testConfig)
 		context := *testData.Context
@@ -200,7 +214,7 @@ func Test_GameByID(t *testing.T) {
 			context.SetParamNames("gameID")
 			context.SetParamValues(tc.gameID)
 
-			midleware := game.GetGameByQueryIDAdmin(func(c echo.Context) error {
+			midleware := game.GetGameByQueryID(func(c echo.Context) error {
 				return nil
 			})
 			err := midleware(context)
@@ -213,10 +227,11 @@ func Test_GameByID(t *testing.T) {
 			err = json.Unmarshal(writer.Body.Bytes(), &resp)
 			assert.NoError(t, err)
 
-			//Test Cases
+			// Test Cases
 			assert.Equal(t, tc.expectedCode, writer.Code)
 			assert.Contains(t, resp.Data, "title")
 			assert.Contains(t, resp.Data, "id")
+			assert.Contains(t, resp.Data, "user_id")
 			assert.Contains(t, resp.Data, "description")
 			assert.Equal(t, tc.expectedResonse.MessageType, resp.MessageType)
 			assert.Equal(t, tc.expectedResonse.Message, resp.Message)
@@ -243,6 +258,7 @@ func Test_CreateGame(t *testing.T) {
 		name            string
 		isBadCase       bool
 		reqBody         map[string]interface{}
+		userID          int
 		expectedCode    int
 		expectedResonse responses.Response
 	}{
@@ -251,6 +267,7 @@ func Test_CreateGame(t *testing.T) {
 			isBadCase:    true,
 			expectedCode: http.StatusBadRequest,
 			reqBody:      map[string]interface{}{},
+			userID:       100,
 			expectedResonse: responses.Response{
 				MessageType: "ERROR",
 				Message:     "All fields are required",
@@ -261,6 +278,7 @@ func Test_CreateGame(t *testing.T) {
 			name:         "Wrong data in reqBody",
 			isBadCase:    true,
 			expectedCode: http.StatusBadRequest,
+			userID:       100,
 			expectedResonse: responses.Response{
 				MessageType: "ERROR",
 				Message:     "Not valid body information",
@@ -275,6 +293,7 @@ func Test_CreateGame(t *testing.T) {
 			name:         "Short title",
 			isBadCase:    true,
 			expectedCode: http.StatusBadRequest,
+			userID:       100,
 			expectedResonse: responses.Response{
 				MessageType: "ERROR",
 				Message:     "Title field must have more than 3 characters",
@@ -289,6 +308,7 @@ func Test_CreateGame(t *testing.T) {
 			name:         "OK",
 			isBadCase:    false,
 			expectedCode: http.StatusCreated,
+			userID:       100,
 			expectedResonse: responses.Response{
 				MessageType: "OK",
 				Message:     "Game created",
@@ -303,10 +323,11 @@ func Test_CreateGame(t *testing.T) {
 
 	for _, tc := range testCases {
 		testConfig := testtools.InitTestConfig{
-			Path:       "/api/v1" + "/games",
-			Method:     http.MethodPost,
-			ReqBody:    tc.reqBody,
-			ApplyToken: true,
+			Path:        "/api/v1" + "/games",
+			Method:      http.MethodPost,
+			ReqBody:     tc.reqBody,
+			ApplyToken:  true,
+			TokenUserID: tc.userID,
 		}
 		testData := testtools.SetGenericTestData(&testConfig)
 		context := *testData.Context
@@ -320,7 +341,7 @@ func Test_CreateGame(t *testing.T) {
 			err = json.Unmarshal(writer.Body.Bytes(), &resp)
 			assert.NoError(t, err)
 
-			//Test Cases
+			// Test Cases
 			if tc.isBadCase {
 				assert.Equal(t, tc.expectedCode, writer.Code)
 				assert.Equal(t, tc.expectedResonse.Data, resp.Data)
@@ -328,6 +349,7 @@ func Test_CreateGame(t *testing.T) {
 				assert.Contains(t, resp.Data, "title")
 				assert.Contains(t, resp.Data, "id")
 				assert.Contains(t, resp.Data, "description")
+				assert.Contains(t, resp.Data, "user_id")
 			}
 			assert.Equal(t, tc.expectedResonse.MessageType, resp.MessageType)
 			assert.Equal(t, tc.expectedResonse.Message, resp.Message)
@@ -337,15 +359,17 @@ func Test_CreateGame(t *testing.T) {
 
 // EditGame endpoint
 func Test_EditGame(t *testing.T) {
-	//Test Constatns
+	// Test Constatns
 	gameID := 1
+	userID := 100
 
-	//TestCases
+	// TestCases
 	environment.SetTestEnvirontment()
 	testCases := []struct {
 		name            string
 		isBadCase       bool
 		gameID          string
+		userID          int
 		reqBody         map[string]interface{}
 		expectedCode    int
 		expectedResonse responses.Response
@@ -353,6 +377,7 @@ func Test_EditGame(t *testing.T) {
 		{
 			name:         "empty field in reqBody",
 			isBadCase:    true,
+			userID:       userID,
 			gameID:       strconv.Itoa(gameID),
 			expectedCode: http.StatusBadRequest,
 			reqBody:      map[string]interface{}{},
@@ -365,6 +390,7 @@ func Test_EditGame(t *testing.T) {
 		{
 			name:         "Wrong data in reqBody",
 			isBadCase:    true,
+			userID:       userID,
 			gameID:       strconv.Itoa(gameID),
 			expectedCode: http.StatusBadRequest,
 			expectedResonse: responses.Response{
@@ -381,6 +407,7 @@ func Test_EditGame(t *testing.T) {
 			name:         "Short title",
 			isBadCase:    true,
 			gameID:       strconv.Itoa(gameID),
+			userID:       userID,
 			expectedCode: http.StatusBadRequest,
 			expectedResonse: responses.Response{
 				MessageType: "ERROR",
@@ -395,12 +422,12 @@ func Test_EditGame(t *testing.T) {
 		{
 			name:         "OK",
 			isBadCase:    false,
+			userID:       userID,
 			gameID:       strconv.Itoa(gameID),
 			expectedCode: http.StatusCreated,
 			expectedResonse: responses.Response{
 				MessageType: "OK",
 				Message:     "Game edited",
-				Data:        dbmodels.Game{},
 			},
 			reqBody: map[string]interface{}{
 				"title":       "valid Title",
@@ -409,7 +436,7 @@ func Test_EditGame(t *testing.T) {
 		},
 	}
 
-	//Mocks
+	// Mocks
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
@@ -424,17 +451,17 @@ func Test_EditGame(t *testing.T) {
 
 	mockUserRepo.
 		EXPECT().
-		GetById(gomock.Eq(gameID), gomock.AssignableToTypeOf(&dbmodels.Game{})).
+		GetGameById(gomock.Eq(userID), gomock.Eq(gameID), gomock.AssignableToTypeOf(&dbmodels.Game{})).
 		Return(&gorm.DB{RowsAffected: 1}).
 		Times(len(testCases))
 
 	for _, tc := range testCases {
 		testConfig := testtools.InitTestConfig{
-			Path:          "/api/v1" + "/games",
-			Method:        http.MethodPut,
-			ReqBody:       tc.reqBody,
-			ApplyToken:    true,
-			TokenUserRole: roles.ADMIN,
+			Path:        "/api/v1" + "/games",
+			Method:      http.MethodPut,
+			ReqBody:     tc.reqBody,
+			ApplyToken:  true,
+			TokenUserID: userID,
 		}
 		testData := testtools.SetGenericTestData(&testConfig)
 		context := *testData.Context
@@ -445,7 +472,7 @@ func Test_EditGame(t *testing.T) {
 			context.SetParamNames("gameID")
 			context.SetParamValues(tc.gameID)
 
-			midleware := game.GetGameByQueryIDAdmin(func(c echo.Context) error {
+			midleware := game.GetGameByQueryID(func(c echo.Context) error {
 				return nil
 			})
 			err := midleware(context)
@@ -458,7 +485,7 @@ func Test_EditGame(t *testing.T) {
 			err = json.Unmarshal(writer.Body.Bytes(), &resp)
 			assert.NoError(t, err)
 
-			//Test Cases
+			// Test Cases
 			if tc.isBadCase {
 				assert.Equal(t, tc.expectedCode, writer.Code)
 				assert.Equal(t, tc.expectedResonse.Data, resp.Data)
@@ -466,6 +493,7 @@ func Test_EditGame(t *testing.T) {
 				assert.Contains(t, resp.Data, "title")
 				assert.Contains(t, resp.Data, "description")
 				assert.Contains(t, resp.Data, "id")
+				assert.Contains(t, resp.Data, "user_id")
 			}
 			assert.Equal(t, tc.expectedResonse.MessageType, resp.MessageType)
 			assert.Equal(t, tc.expectedResonse.Message, resp.Message)
@@ -475,17 +503,60 @@ func Test_EditGame(t *testing.T) {
 
 // DeleteGame endpoint
 func Test_DeletGame(t *testing.T) {
+	foundID := 1
+	userID := 100
+	NouserID := 200
+
+	environment.SetTestEnvirontment()
+	testCases := []struct {
+		name            string
+		gameID          string
+		userID          int
+		isAuthor        bool
+		expectedCode    int
+		expectedResonse responses.Response
+	}{
+		{
+			name:         "Must delete the game",
+			gameID:       strconv.Itoa(foundID),
+			userID:       userID,
+			isAuthor:     true,
+			expectedCode: http.StatusOK,
+			expectedResonse: responses.Response{
+				MessageType: "OK",
+				Message:     "Game Deleted",
+			},
+		},
+		{
+			name:         "Must not find the game",
+			gameID:       strconv.Itoa(foundID),
+			userID:       NouserID,
+			isAuthor:     false,
+			expectedCode: http.StatusNotFound,
+			expectedResonse: responses.Response{
+				MessageType: "ERROR",
+				Message:     fmt.Sprintf("Game %d not found", foundID),
+				Data:        nil,
+			},
+		},
+	}
+
 	ctrl := gomock.NewController(t)
 	defer ctrl.Finish()
 
 	mockUserRepo := mock_models.NewMockIstorage(ctrl)
 	game := gameshandlers.NewGame(mockUserRepo)
 
-	foundID := 1
 	mockUserRepo.
 		EXPECT().
-		GetById(gomock.Eq(foundID), gomock.AssignableToTypeOf(&dbmodels.Game{})).
+		GetGameById(gomock.Eq(userID), gomock.Eq(foundID), gomock.AssignableToTypeOf(&dbmodels.Game{})).
 		Return(&gorm.DB{RowsAffected: 1}).
+		Times(1)
+
+	mockUserRepo.
+		EXPECT().
+		GetGameById(gomock.Eq(NouserID), gomock.Eq(foundID), gomock.AssignableToTypeOf(&dbmodels.Game{})).
+		Return(&gorm.DB{RowsAffected: 0}).
 		Times(1)
 
 	mockUserRepo.
@@ -494,34 +565,13 @@ func Test_DeletGame(t *testing.T) {
 		Return(nil).
 		Times(1)
 
-	environment.SetTestEnvirontment()
-	testCases := []struct {
-		name            string
-		gameID          string
-		userRole        string
-		expectedCode    int
-		expectedResonse responses.Response
-	}{
-		{
-			name:         "Must delete the game",
-			gameID:       strconv.Itoa(foundID),
-			userRole:     roles.ADMIN,
-			expectedCode: http.StatusOK,
-			expectedResonse: responses.Response{
-				MessageType: "OK",
-				Message:     "Game Deleted",
-				Data:        nil,
-			},
-		},
-	}
-
 	for _, tc := range testCases {
 		testConfig := testtools.InitTestConfig{
-			Path:          "/api/v1" + "/games",
-			Method:        http.MethodGet,
-			ReqBody:       nil,
-			ApplyToken:    true,
-			TokenUserRole: tc.userRole,
+			Path:        "/api/v1" + "/games",
+			Method:      http.MethodGet,
+			ReqBody:     nil,
+			ApplyToken:  true,
+			TokenUserID: tc.userID,
 		}
 		testData := testtools.SetGenericTestData(&testConfig)
 		context := *testData.Context
@@ -532,24 +582,31 @@ func Test_DeletGame(t *testing.T) {
 			context.SetParamNames("gameID")
 			context.SetParamValues(tc.gameID)
 
-			midleware := game.GetGameByQueryIDAdmin(func(c echo.Context) error {
+			midleware := game.GetGameByQueryID(func(c echo.Context) error {
 				return nil
 			})
 			err := midleware(context)
 			assert.NoError(t, err)
 
-			err = game.DeleteGame(context)
-			assert.NoError(t, err)
+			if tc.isAuthor {
+				err = game.DeleteGame(context)
+				assert.NoError(t, err)
+			}
 
 			resp := responses.Response{}
 			err = json.Unmarshal(writer.Body.Bytes(), &resp)
 			assert.NoError(t, err)
 
-			//Test Cases
+			// Test Cases
+			if !tc.isAuthor {
+				assert.Equal(t, tc.expectedResonse.Data, resp.Data)
+			} else {
+				assert.Contains(t, resp.Data, "title")
+				assert.Contains(t, resp.Data, "description")
+				assert.Contains(t, resp.Data, "id")
+				assert.Contains(t, resp.Data, "user_id")
+			}
 			assert.Equal(t, tc.expectedCode, writer.Code)
-			assert.Contains(t, resp.Data, "title")
-			assert.Contains(t, resp.Data, "id")
-			assert.Contains(t, resp.Data, "description")
 			assert.Equal(t, tc.expectedResonse.MessageType, resp.MessageType)
 			assert.Equal(t, tc.expectedResonse.Message, resp.Message)
 		})
